+++ src/cashbook/cashbook.body/data/Transaction.cs
@@ -8,8 +8,8 @@ namespace cashbook.body.data
 	public class Transaction {
 		public TransactionTypes Type;
 		public DateTime TransactionDate;
-		public double Amount;
-		public double Value { get { return Type == TransactionTypes.Deposit ? Amount : -Amount; } }
+		public decimal Amount;
+		public decimal Value { get { return Type == TransactionTypes.Deposit ? Amount : -Amount; } }
 		public string Description;
 	}	
 }
