+++ src/cashbook/cashbook.contracts/data/Balance.cs
@@ -5,6 +5,6 @@ namespace cashbook.contracts.data
 {
 	public class Balance {
 		public DateTime Month;
-		public double Value;
+		public decimal Value;
 	}
 }
